@@ -2,7 +2,7 @@
   "name": "testshot",
   "version": "0.1.0",
   "description": "testshot",
-  "main": "./dist/app.js",
+  "main": "./lib/index.js",
   "author": "Eduard Tsech",
   "homepage": "https://github.com/toptal/testshot",
   "repository": {
@@ -14,35 +14,25 @@
   },
   "dependencies": {
     "body-parser": "^1.15.2",
-    "classnames": "^2.2.5",
     "cors": "^2.8.1",
-    "enzyme": "^2.6.0",
-    "enzyme-to-json": "1.4.5",
     "express": "^4.14.0",
-    "express-cors": "0.0.3",
-    "fs": "0.0.1-security",
     "html-differ": "^1.3.4",
-    "html-minifier": "^3.2.3",
     "jfs": "^0.2.6",
-    "jsondiffpatch": "^0.2.4",
-    "jsx-to-string": "^1.0.0",
     "lodash": "^4.17.2",
-    "prettydiff": "^99.0.1",
-    "react-addons-test-utils": "^15.4.1",
-    "react-test-renderer": "^15.4.1",
     "styled-components": "^1.3.1",
     "styled-mixin": "^0.1.2"
   },
   "devDependencies": {
     "autoprefixer": "^6.7.0",
     "babel": "^6.5.2",
+    "babel-cli": "^6.22.2",
     "babel-core": "^6.21.0",
     "babel-eslint": "^4.1.3",
     "babel-jest": "^18.0.0",
     "babel-loader": "^6.2.10",
     "babel-plugin-lodash": "^3.2.11",
     "babel-polyfill": "^6.20.0",
-    "babel-preset-es2015": "^6.18.0",
+    "babel-preset-es2015": "^6.22.0",
     "babel-preset-react": "^6.22.0",
     "babel-preset-stage-0": "^6.22.0",
     "css-loader": "^0.26.1",
@@ -50,11 +40,9 @@
     "eslint": "3.4.0",
     "eslint-plugin-mocha": "^4.8.0",
     "eslint-plugin-react": "^6.9.0",
-    "estraverse-fb": "^1.3.1",
     "json-loader": "^0.5.4",
     "postcss": "^5.2.11",
     "postcss-loader": "^1.2.2",
-    "sass-loader": "^4.1.1",
     "standard": "^8.6.0",
     "standard-jsx": "^8.0.0",
     "standard-react": "^4.0.0",
@@ -64,14 +52,13 @@
     "webpack-hot-middleware": "^2.15.0"
   },
   "peerDependencies": {
-    "react": "^0.14.0",
-    "react-dom": "^0.14.0 || ^15.0.0-0"
+    "react": ">=15 <16",
+    "react-dom": ">=15 <16"
   },
   "scripts": {
     "dev": "NODE_ENV=development node devServer.js",
     "dev-server": "TESTSHOT_CONFIG='./example/testshot.config.json' node bin/server.js",
-    "build": "NODE_ENV=production webpack",
-    "postinstall": "NODE_ENV=production webpack",
+    "build": "babel src --source-root src --out-dir lib --ignore test.js",
     "release": "TODO...",
     "lint": "eslint ./; true",
     "test": "echo \"no tests yet\" && exit 0"
