@@ -1,9 +1,10 @@
-// XXX: Main entry point
-// Exposes public methods of the application
-import init from 'src/init'
-import {scenario} from 'src/Testshot'
+// Main entry point, exposes the public API
+
+import init from './init'
+import TestshotComponent, {scenario} from './Testshot'
 
 module.exports = {
   init,
-  scenario
+  scenario,
+  Testshot: TestshotComponent
 }
